requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==0.19.2
dnspython==2.2.1
fastapi
motor
beanie
python-jose[cryptography]
passlib[bcrypt]
redis
websockets
pytest